rockchip: rk3506: Enable use of SPL_OPTEE_IMAGE
Vendor OP-TEE blob provides PSCI that is used to enable CPU cores on
RK3506. The OP-TEE blob is expected to be loaded and executed from
0x1000 and uses memory in the 0-1 MiB area of DRAM.

Add initial support for loading OP-TEE firmware on RK3506. Reserve the
first 1 MiB of DRAM for OP-TEE and use ROCKCHIP_SPL_RESERVE_IRAM to add
padding to the start of SPL text area to allow loading OP-TEE to 0x1000.

Signed-off-by: Austin Lane <vidplace7@gmail.com>